+++ XenAdmin/TabPages/ManageUpdatesPage.cs
@@ -1009,33 +1009,6 @@ namespace XenAdmin.TabPages
             labelProgress.MaximumSize = new Size(tableLayoutPanel3.Width - 60, tableLayoutPanel3.Size.Height);
         }
 
-        public class NewVersionPriorityAlertComparer : IComparer<Alert>
-        {
-            public int Compare(Alert alert1, Alert alert2)
-            {
-                if (alert1 == null || alert2 == null)
-                    return 0;
 
-                int sortResult = 0;
-
-                if (IsVersionOrVersionUpdateAlert(alert1) && !IsVersionOrVersionUpdateAlert(alert2))
-                    sortResult = 1;
-
-                if (!IsVersionOrVersionUpdateAlert(alert1) && IsVersionOrVersionUpdateAlert(alert2))
-                    sortResult = -1;
-
-                if (sortResult == 0)
-                    sortResult = Alert.CompareOnDate(alert1, alert2);
-
-                return -sortResult;
-            }
-
-            private bool IsVersionOrVersionUpdateAlert(Alert alert)
-            {
-                return alert is XenServerPatchAlert && (alert as XenServerPatchAlert).NewServerVersion != null
-                    || alert is XenServerVersionAlert
-                    || alert is XenCenterUpdateAlert;
-            }
-        }
     }
 }
